UGENE-7741 fix tests
