+++ src/corelibs/U2Algorithm/src/temperature/TempCalcFactory.h
@@ -72,7 +72,7 @@ public:
 
     const QString& getId() const;
 
-private:
+protected:
     const QString id;
 };
