+++ src/corelibs/U2Algorithm/src/temperature/methods/RoughTempCalc.cpp
