+++ src/corelibs/U2View/src/temperature/BaseTempCalcWidget.cpp
