+++ src/corelibs/U2View/src/temperature/TempCalcDialog.cpp
