+++ src/corelibs/U2View/src/temperature/TempCalcWidget.cpp
