+++ src/include/U2Algorithm/RoughTempCalcCmdFactory.h
@@ -1,0 +1,1 @@
+#include "../../corelibs/U2Algorithm/src/temperature/methods/RoughTempCalcCmdFactory.h"
