+++ src/include/U2View/RoughTempCalcFactory.h
@@ -1,1 +1,1 @@
-#include "../../corelibs/U2View/src/ov_sequence/temperature/methods/RoughTempCalcFactory.h"
+#include "../../corelibs/U2View/src/temperature/methods/RoughTempCalcFactory.h"
